requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    make upgrade
#
asgiref==3.6.0
    # via
    #   -r requirements/base.txt
    #   django
attrs==22.2.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
    #   pytest
certifi==2022.12.7
    # via
    #   -r requirements/base.txt
    #   requests
charset-normalizer==3.1.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via code-annotations
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==7.2.2
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   openedx-events
    #   openedx-filters
django-crum==0.7.9
    # via -r requirements/base.txt
edx-opaque-keys[django]==2.3.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
exceptiongroup==1.1.1
    # via pytest
fastavro==1.7.3
    # via
    #   -r requirements/base.txt
    #   openedx-events
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via code-annotations
markupsafe==2.1.2
    # via jinja2
openedx-events==7.0.0
    # via -r requirements/base.txt
openedx-filters==1.2.0
    # via -r requirements/base.txt
packaging==23.0
    # via pytest
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.0.0
    # via pytest
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pytest==7.2.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-slugify==8.0.1
    # via code-annotations
pytz==2022.7.1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==6.0
    # via code-annotations
requests==2.28.2
    # via -r requirements/base.txt
sqlparse==0.4.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.0.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
urllib3==1.26.15
    # via
    #   -r requirements/base.txt
    #   requests
